fixed build system
