+++ extension/src-webview/tsconfig.json
@@ -1,28 +1,14 @@
 {
+  "extends": "../tsconfig.json",
   "compilerOptions": {
-      "module": "CommonJS",
-      "target": "ES2017",
-      "outDir": "pack",
-      "lib": ["es6", "dom"],
-      "jsx": "react",
-      "jsxFactory": "jsx",
-      "sourceMap": true,
-      "rootDir": "src",
-      "strict": true, // enable all strict type-checking options
-      "resolveJsonModule": true, // Allow importing JSON files
-      "moduleResolution": "node",
-      "esModuleInterop": true,
-      /* Additional Checks */
-      "strictPropertyInitialization": false, // No need to initialize every variable
-      "declaration": false,
-      "noImplicitAny": false,
-      "noEmitOnError": true,
-      "noImplicitThis": true,
-      "strictNullChecks": true,
-      "experimentalDecorators": true,
-      "emitDecoratorMetadata": true,
-      "skipLibCheck": true,
-  },
-  // "include": ["src"],
-  "exclude": ["node_modules", ".vscode-test", "pack"]
+    "target": "ES2017",
+    "lib": [
+        "es2017",
+        "dom"
+    ],
+    "rootDir": ".",
+    "module": "commonjs",
+    "moduleResolution": "node",
+    "jsx": "react",
+}
 }
