+++ extension/src/context-table-panel.ts
@@ -17,8 +17,8 @@
 
 import { TableWebview } from "@kieler/table-webview/lib/table-webview";
 import * as vscode from "vscode";
-import { SendContextTableDataAction } from "../src-context-table/actions";
-import { ContextTableData } from "../src-context-table/utils";
+import { SendContextTableDataAction } from "./actions";
+import { ContextTableData } from "../src-context-table/src/utils-classes";
 
 export class ContextTablePanel extends TableWebview {
     constructor(identifier: string, localResourceRoots: vscode.Uri[], scriptUri: vscode.Uri) {
