+++ extension/tsconfig.json
@@ -1,21 +1,13 @@
 {
   "compilerOptions": {
-      "module": "Node16",
-      "target": "ES2017",
       "outDir": "pack",
-      "lib": ["ESNext", "dom"],
-      "jsx": "react",
-      "jsxFactory": "jsx",
       "sourceMap": true,
-      "rootDir": "src",
       "strict": true, // enable all strict type-checking options
       "resolveJsonModule": true, // Allow importing JSON files
-      "moduleResolution": "Node16",
-      "esModuleInterop": true,
       /* Additional Checks */
       "strictPropertyInitialization": false, // No need to initialize every variable
-      "declaration": false,
-      "noImplicitAny": false,
+      "declaration": true, // Generate corresponding .d.ts file
+      "noImplicitAny": true,
       "noEmitOnError": true,
       "noImplicitThis": true,
       "strictNullChecks": true,
@@ -23,6 +15,6 @@
       "emitDecoratorMetadata": true,
       "skipLibCheck": true,
   },
-  // "include": ["src"],
+  
   "exclude": ["node_modules", ".vscode-test", "pack"]
 }
