+++ package.json
@@ -16,7 +16,7 @@
   },
   "devDependencies": {
     "lerna": "^3.20.2",
-    "typescript": "^4.3.2",
+    "typescript": "^4.9.3",
     "rewire": "^6.0.0",
     "mocha": "^10.2.0",
     "@types/chai": "^4.3.4",
